base class updates

+++ tests/Answer.King.IntegrationTests/Utilities/BaseTestClass.cs
@@ -2,23 +2,34 @@ namespace Answer.King.IntegrationTests.Utilities;
 using System;
 using System.IO;
 using System.Net.Http;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 public class BaseTestClass : IDisposable
 {
     public HttpClient Client;
 
-    [OneTimeSetUp]
+    public string testDb { get; set; } = $"Answer.King.{Guid.NewGuid()}.db";
+
+    [SetUp]
     public void Setup()
     {
-        File.Delete(Path.Combine(TestContext.CurrentContext.TestDirectory, "db/Answer.King.db"));
-        File.Delete(Path.Combine(TestContext.CurrentContext.TestDirectory, "db/Answer.King-log.db"));
-        this.Client = new WebApplicationFactory<Program>().CreateClient();
+        this.Client = new WebApplicationFactory<Program>().WithWebHostBuilder(builder =>
+        {
+            builder.UseSetting("ConnectionStrings:AnswerKing", $"filename=db/{this.testDb};Connection=Shared");
+        }).CreateClient();
     }
 
-    [OneTimeTearDown]
+    [TearDown]
     public void TearDown()
     {
+        if (Directory.Exists(Path.Combine(TestContext.CurrentContext.TestDirectory, "db")))
+        {
+            File.Delete(Path.Combine(TestContext.CurrentContext.TestDirectory, $"db/{testDb}"));
+            File.Delete(Path.Combine(TestContext.CurrentContext.TestDirectory, "db/Answer.King-log.db"));
+        }
         this.Client.Dispose();
     }
